refactor(store): closer to endo
